requirements.txt
yt-dlp
openai-whisper
torch
transformers
pytube